Refactor: STT
