+++ src/speech/STT.py
@@ -6,9 +6,9 @@ import wave
 # from CLOVA_STT import stt_clova
 import google.cloud.speech_v1p1beta1 as speech
 import google.cloud.storage as storage
-from convert import convert_audio_files, convert_text_data
+from audio import convert_audio_files, convert_text_data
 from evaluate import remove_correct, evaluate_SER
-from tuning_STT.expand import remove_duplication
+from speech.util import remove_duplication
 from common.info import open_dialog
 import sys
