+++ src/speech/util.py
